+++ library/core/edhoc_message_4.c
@@ -25,6 +25,7 @@ LOG_MODULE_DECLARE(libedhoc, CONFIG_LIBEDHOC_LOG_LEVEL);
 #include "edhoc_context_internal.h"
 #include "edhoc_values_internal.h"
 #include "edhoc_macros_internal.h"
+#include "edhoc_cbor_internal.h"
 #include "edhoc_common_internal.h"
 #include "edhoc_backend_log.h"
 #include "edhoc_backend_memory.h"
@@ -196,9 +197,9 @@ STATIC int compute_plaintext_4_len(const struct edhoc_context *ctx,
 	size_t len = 0;
 
 	for (size_t i = 0; i < ctx->ead.count; ++i) {
-		len += edhoc_cbor_int_length(ctx->ead.token[i].label);
+		len += edhoc_cbor_int_head_length(ctx->ead.token[i].label);
 		len += ctx->ead.token[i].value.length;
-		len += edhoc_cbor_bstr_header_length(
+		len += edhoc_cbor_bstr_head_length(
 			ctx->ead.token[i].value.length);
 	}
 
@@ -256,10 +257,10 @@ STATIC size_t compute_aad_4_len(const struct edhoc_context *ctx)
 	size_t len = 0;
 
 	len += sizeof("Encrypt0") +
-	       edhoc_cbor_tstr_header_length(sizeof("Encrypt0"));
-	len += edhoc_cbor_bstr_header_length(0);
+	       edhoc_cbor_tstr_head_length(sizeof("Encrypt0"));
+	len += edhoc_cbor_bstr_head_length(0);
 	len += ctx->state.th.length +
-	       edhoc_cbor_bstr_header_length(ctx->state.th.length);
+	       edhoc_cbor_bstr_head_length(ctx->state.th.length);
 
 	return len;
 }
@@ -289,10 +290,10 @@ STATIC int compute_key_iv_aad_4(struct edhoc_context *ctx, uint8_t *iv,
 
 	/* Calculate struct info cbor overhead. */
 	size_t len = 0;
-	len += edhoc_cbor_int_length(EDHOC_EXTRACT_PRK_INFO_LABEL_IV_4);
+	len += edhoc_cbor_int_head_length(EDHOC_EXTRACT_PRK_INFO_LABEL_IV_4);
 	len += ctx->state.th.length +
-	       edhoc_cbor_bstr_header_length(ctx->state.th.length);
-	len += edhoc_cbor_int_length((int32_t)csuite->aead_key_length);
+	       edhoc_cbor_bstr_head_length(ctx->state.th.length);
+	len += edhoc_cbor_int_head_length((int32_t)csuite->aead_key_length);
 
 	EDHOC_MEM_ALLOC(uint8_t, info, len);
 	if (NULL == info) {